Prevent serializing 9001=Celcius instead of 0=LocaleDefault
Apparently there's issues with the WeatherPlugin singleton models
when you switch to Units => another config => back to Units.

For now I've disabled the ComboBoxes as no changes are serialized
anyway since the models aren't "populated". The ComboBoxes show
"Celcius", which leads me to believe the "tab" was "hibernated"
instead of destroyed and recreated.